refactor: 로그인 로딩 페이지 경로 변경
- splash 폴더 내부로 이동
